+++ CHANGELOG.md
@@ -65,6 +65,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `jj util diff <path1> <path2>` to compare files on disk.
 
+* `jj workspace add` now has `--adopt` for instantaneous clones of repos when
+  used with a CoW filesystem. For example, by taking a snapshot with
+  `btrfs subvolume snapshot . ../workspace`, removing `../workspace/.jj`, and
+  adopting it with `jj workspace add --adopt ../workspace`.
 ### Fixed bugs
 
 * On Windows, `jj` no longer hangs when a subprocess needs to prompt the user,